fix(deepseek): export resolveThinkingProfile from provider policy surface
Extract DeepSeek V4 thinking profile logic into a shared module
(thinking-profile.ts) and wire it through the bundled provider policy
surface (provider-policy-api.ts). This allows the fallback path in
src/plugins/provider-thinking.ts to resolve xhigh and max thinking
levels for native DeepSeek V4 models via resolveBundledProviderPolicySurface.

Previously, resolveThinkingProfile was only registered at runtime
(index.ts) but not exported from the lightweight policy surface,
causing the Control UI /think picker to show only the base five levels
(off/minimal/low/medium/high) when hitting the fallback path.

Fixes #77139
