+++ extensions/deepseek/index.ts
@@ -1,27 +1,12 @@
-import type { ProviderThinkingProfile } from "openclaw/plugin-sdk/plugin-entry";
 import { readConfiguredProviderCatalogEntries } from "openclaw/plugin-sdk/provider-catalog-shared";
 import { defineSingleProviderPluginEntry } from "openclaw/plugin-sdk/provider-entry";
 import { buildProviderReplayFamilyHooks } from "openclaw/plugin-sdk/provider-model-shared";
-import { isDeepSeekV4ModelId } from "./models.js";
 import { applyDeepSeekConfig, DEEPSEEK_DEFAULT_MODEL_REF } from "./onboard.js";
 import { buildDeepSeekProvider } from "./provider-catalog.js";
 import { createDeepSeekV4ThinkingWrapper } from "./stream.js";
+import { resolveDeepSeekV4ThinkingProfile } from "./thinking-profile.js";
 
 const PROVIDER_ID = "deepseek";
-const V4_THINKING_LEVEL_IDS = ["off", "minimal", "low", "medium", "high", "xhigh", "max"] as const;
-
-function buildDeepSeekV4ThinkingLevel(id: (typeof V4_THINKING_LEVEL_IDS)[number]) {
-  return { id };
-}
-
-const DEEPSEEK_V4_THINKING_PROFILE = {
-  levels: V4_THINKING_LEVEL_IDS.map(buildDeepSeekV4ThinkingLevel),
-  defaultLevel: "high",
-} satisfies ProviderThinkingProfile;
-
-function resolveDeepSeekV4ThinkingProfile(modelId: string): ProviderThinkingProfile | undefined {
-  return isDeepSeekV4ModelId(modelId) ? DEEPSEEK_V4_THINKING_PROFILE : undefined;
-}
 
 export default defineSingleProviderPluginEntry({
   id: PROVIDER_ID,
